file structure : renamed component folder to components and moved data folder out of components
