+++ src/components/CalculateVariables.jsx
