+++ src/components/CharacterObjects.js
