+++ src/components/Notepad.jsx
