+++ src/components/PlayerLost.jsx
