+++ src/components/gameMechanics.js
