+++ src/components/styles/CharacterMover.css
