+++ src/components/styles/Notepad.css
