+++ src/components/styles/Timer.css
