+++ src/components/styles/WelcomePopUp.css
