+++ src/data/Characters/9-Richard-Albright.json
