+++ src/data/Gen-dialog/OpeningScene.json
